Encapsulate ShapePlugin into its own module
